+++ internal/provider/model_catalystcenter_lan_automation.go
@@ -31,17 +31,20 @@ import (
 
 // Section below is generated&owned by "gen/generator.go". //template:begin types
 type LANAutomation struct {
-	Id                                types.String           `tfsdk:"id"`
-	DiscoveredDeviceSiteNameHierarchy types.String           `tfsdk:"discovered_device_site_name_hierarchy"`
-	PrimaryDeviceManagementIpAddress  types.String           `tfsdk:"primary_device_management_ip_address"`
-	PeerDeviceManagementIpAddress     types.String           `tfsdk:"peer_device_management_ip_address"`
-	PrimaryDeviceInterfaceNames       types.Set              `tfsdk:"primary_device_interface_names"`
-	IpPools                           []LANAutomationIpPools `tfsdk:"ip_pools"`
-	MulticastEnabled                  types.Bool             `tfsdk:"multicast_enabled"`
-	HostNamePrefix                    types.String           `tfsdk:"host_name_prefix"`
-	HostNameFileId                    types.String           `tfsdk:"host_name_file_id"`
-	IsisDomainPassword                types.String           `tfsdk:"isis_domain_password"`
-	RedistributeIsisToBgp             types.Bool             `tfsdk:"redistribute_isis_to_bgp"`
+	Id                                types.String                    `tfsdk:"id"`
+	DiscoveredDeviceSiteNameHierarchy types.String                    `tfsdk:"discovered_device_site_name_hierarchy"`
+	PrimaryDeviceManagementIpAddress  types.String                    `tfsdk:"primary_device_management_ip_address"`
+	PeerDeviceManagementIpAddress     types.String                    `tfsdk:"peer_device_management_ip_address"`
+	PrimaryDeviceInterfaceNames       types.Set                       `tfsdk:"primary_device_interface_names"`
+	IpPools                           []LANAutomationIpPools          `tfsdk:"ip_pools"`
+	MulticastEnabled                  types.Bool                      `tfsdk:"multicast_enabled"`
+	HostNamePrefix                    types.String                    `tfsdk:"host_name_prefix"`
+	HostNameFileId                    types.String                    `tfsdk:"host_name_file_id"`
+	IsisDomainPassword                types.String                    `tfsdk:"isis_domain_password"`
+	RedistributeIsisToBgp             types.Bool                      `tfsdk:"redistribute_isis_to_bgp"`
+	DiscoveryLevel                    types.Int64                     `tfsdk:"discovery_level"`
+	DiscoveryTimeout                  types.Int64                     `tfsdk:"discovery_timeout"`
+	DiscoveryDevices                  []LANAutomationDiscoveryDevices `tfsdk:"discovery_devices"`
 }
 
 type LANAutomationIpPools struct {
@@ -49,16 +52,27 @@ type LANAutomationIpPools struct {
 	IpPoolRole types.String `tfsdk:"ip_pool_role"`
 }
 
+type LANAutomationDiscoveryDevices struct {
+	DeviceSerialNumber        types.String `tfsdk:"device_serial_number"`
+	DeviceHostName            types.String `tfsdk:"device_host_name"`
+	DeviceSiteNameHierarchy   types.String `tfsdk:"device_site_name_hierarchy"`
+	DeviceManagementIpAddress types.String `tfsdk:"device_management_ip_address"`
+}
+
 // End of section. //template:end types
 
 // Section below is generated&owned by "gen/generator.go". //template:begin getPath
 func (data LANAutomation) getPath() string {
-	return "/dna/intent/api/v1/lan-automation"
+	return "/dna/intent/api/v2/lan-automation"
 }
 
 // End of section. //template:end getPath
 
 // Section below is generated&owned by "gen/generator.go". //template:begin getPathDelete
+
+func (data LANAutomation) getPathDelete() string {
+	return "/dna/intent/api/v1/lan-automation"
+}
 
 // End of section. //template:end getPathDelete
 
@@ -112,6 +126,31 @@ func (data LANAutomation) toBody(ctx context.Context, state LANAutomation) strin
 	if !data.RedistributeIsisToBgp.IsNull() {
 		body, _ = sjson.Set(body, "0.redistributeIsisToBgp", data.RedistributeIsisToBgp.ValueBool())
 	}
+	if !data.DiscoveryLevel.IsNull() {
+		body, _ = sjson.Set(body, "0.discoveryLevel", data.DiscoveryLevel.ValueInt64())
+	}
+	if !data.DiscoveryTimeout.IsNull() {
+		body, _ = sjson.Set(body, "0.discoveryTimeout", data.DiscoveryTimeout.ValueInt64())
+	}
+	if len(data.DiscoveryDevices) > 0 {
+		body, _ = sjson.Set(body, "0.discoveryDevices", []interface{}{})
+		for _, item := range data.DiscoveryDevices {
+			itemBody := ""
+			if !item.DeviceSerialNumber.IsNull() {
+				itemBody, _ = sjson.Set(itemBody, "deviceSerialNumber", item.DeviceSerialNumber.ValueString())
+			}
+			if !item.DeviceHostName.IsNull() {
+				itemBody, _ = sjson.Set(itemBody, "deviceHostName", item.DeviceHostName.ValueString())
+			}
+			if !item.DeviceSiteNameHierarchy.IsNull() {
+				itemBody, _ = sjson.Set(itemBody, "deviceSiteNameHierarchy", item.DeviceSiteNameHierarchy.ValueString())
+			}
+			if !item.DeviceManagementIpAddress.IsNull() {
+				itemBody, _ = sjson.Set(itemBody, "deviceManagementIPAddress", item.DeviceManagementIpAddress.ValueString())
+			}
+			body, _ = sjson.SetRaw(body, "0.discoveryDevices.-1", itemBody)
+		}
+	}
 	return body
 }
 
@@ -157,7 +196,7 @@ func (data *LANAutomation) fromBody(ctx context.Context, res gjson.Result) {
 			return true
 		})
 	}
-	if value := res.Get("response.0.multicatEnabled"); value.Exists() {
+	if value := res.Get("response.0.multicastEnabled"); value.Exists() {
 		data.MulticastEnabled = types.BoolValue(value.Bool())
 	} else {
 		data.MulticastEnabled = types.BoolNull()
@@ -172,15 +211,48 @@ func (data *LANAutomation) fromBody(ctx context.Context, res gjson.Result) {
 	} else {
 		data.HostNameFileId = types.StringNull()
 	}
-	if value := res.Get("response.0.isisDomainPwd"); value.Exists() {
-		data.IsisDomainPassword = types.StringValue(value.String())
-	} else {
-		data.IsisDomainPassword = types.StringNull()
-	}
 	if value := res.Get("response.0.redistributeIsisToBgp"); value.Exists() {
 		data.RedistributeIsisToBgp = types.BoolValue(value.Bool())
 	} else {
 		data.RedistributeIsisToBgp = types.BoolNull()
+	}
+	if value := res.Get("response.0.discoveryLevel"); value.Exists() {
+		data.DiscoveryLevel = types.Int64Value(value.Int())
+	} else {
+		data.DiscoveryLevel = types.Int64Value(2)
+	}
+	if value := res.Get("response.0.discoveryTimeout"); value.Exists() {
+		data.DiscoveryTimeout = types.Int64Value(value.Int())
+	} else {
+		data.DiscoveryTimeout = types.Int64Null()
+	}
+	if value := res.Get("response.0.discoveryDevices"); value.Exists() && len(value.Array()) > 0 {
+		data.DiscoveryDevices = make([]LANAutomationDiscoveryDevices, 0)
+		value.ForEach(func(k, v gjson.Result) bool {
+			item := LANAutomationDiscoveryDevices{}
+			if cValue := v.Get("deviceSerialNumber"); cValue.Exists() {
+				item.DeviceSerialNumber = types.StringValue(cValue.String())
+			} else {
+				item.DeviceSerialNumber = types.StringNull()
+			}
+			if cValue := v.Get("deviceHostName"); cValue.Exists() {
+				item.DeviceHostName = types.StringValue(cValue.String())
+			} else {
+				item.DeviceHostName = types.StringNull()
+			}
+			if cValue := v.Get("deviceSiteNameHierarchy"); cValue.Exists() {
+				item.DeviceSiteNameHierarchy = types.StringValue(cValue.String())
+			} else {
+				item.DeviceSiteNameHierarchy = types.StringNull()
+			}
+			if cValue := v.Get("deviceManagementIPAddress"); cValue.Exists() {
+				item.DeviceManagementIpAddress = types.StringValue(cValue.String())
+			} else {
+				item.DeviceManagementIpAddress = types.StringNull()
+			}
+			data.DiscoveryDevices = append(data.DiscoveryDevices, item)
+			return true
+		})
 	}
 }
 
@@ -242,7 +314,7 @@ func (data *LANAutomation) updateFromBody(ctx context.Context, res gjson.Result)
 			data.IpPools[i].IpPoolRole = types.StringNull()
 		}
 	}
-	if value := res.Get("response.0.multicatEnabled"); value.Exists() && !data.MulticastEnabled.IsNull() {
+	if value := res.Get("response.0.multicastEnabled"); value.Exists() && !data.MulticastEnabled.IsNull() {
 		data.MulticastEnabled = types.BoolValue(value.Bool())
 	} else {
 		data.MulticastEnabled = types.BoolNull()
@@ -257,15 +329,64 @@ func (data *LANAutomation) updateFromBody(ctx context.Context, res gjson.Result)
 	} else {
 		data.HostNameFileId = types.StringNull()
 	}
-	if value := res.Get("response.0.isisDomainPwd"); value.Exists() && !data.IsisDomainPassword.IsNull() {
-		data.IsisDomainPassword = types.StringValue(value.String())
-	} else {
-		data.IsisDomainPassword = types.StringNull()
-	}
 	if value := res.Get("response.0.redistributeIsisToBgp"); value.Exists() && !data.RedistributeIsisToBgp.IsNull() {
 		data.RedistributeIsisToBgp = types.BoolValue(value.Bool())
 	} else {
 		data.RedistributeIsisToBgp = types.BoolNull()
+	}
+	if value := res.Get("response.0.discoveryLevel"); value.Exists() && !data.DiscoveryLevel.IsNull() {
+		data.DiscoveryLevel = types.Int64Value(value.Int())
+	} else if data.DiscoveryLevel.ValueInt64() != 2 {
+		data.DiscoveryLevel = types.Int64Null()
+	}
+	if value := res.Get("response.0.discoveryTimeout"); value.Exists() && !data.DiscoveryTimeout.IsNull() {
+		data.DiscoveryTimeout = types.Int64Value(value.Int())
+	} else {
+		data.DiscoveryTimeout = types.Int64Null()
+	}
+	for i := range data.DiscoveryDevices {
+		keys := [...]string{"deviceSerialNumber"}
+		keyValues := [...]string{data.DiscoveryDevices[i].DeviceSerialNumber.ValueString()}
+
+		var r gjson.Result
+		res.Get("response.0.discoveryDevices").ForEach(
+			func(_, v gjson.Result) bool {
+				found := false
+				for ik := range keys {
+					if v.Get(keys[ik]).String() == keyValues[ik] {
+						found = true
+						continue
+					}
+					found = false
+					break
+				}
+				if found {
+					r = v
+					return false
+				}
+				return true
+			},
+		)
+		if value := r.Get("deviceSerialNumber"); value.Exists() && !data.DiscoveryDevices[i].DeviceSerialNumber.IsNull() {
+			data.DiscoveryDevices[i].DeviceSerialNumber = types.StringValue(value.String())
+		} else {
+			data.DiscoveryDevices[i].DeviceSerialNumber = types.StringNull()
+		}
+		if value := r.Get("deviceHostName"); value.Exists() && !data.DiscoveryDevices[i].DeviceHostName.IsNull() {
+			data.DiscoveryDevices[i].DeviceHostName = types.StringValue(value.String())
+		} else {
+			data.DiscoveryDevices[i].DeviceHostName = types.StringNull()
+		}
+		if value := r.Get("deviceSiteNameHierarchy"); value.Exists() && !data.DiscoveryDevices[i].DeviceSiteNameHierarchy.IsNull() {
+			data.DiscoveryDevices[i].DeviceSiteNameHierarchy = types.StringValue(value.String())
+		} else {
+			data.DiscoveryDevices[i].DeviceSiteNameHierarchy = types.StringNull()
+		}
+		if value := r.Get("deviceManagementIPAddress"); value.Exists() && !data.DiscoveryDevices[i].DeviceManagementIpAddress.IsNull() {
+			data.DiscoveryDevices[i].DeviceManagementIpAddress = types.StringValue(value.String())
+		} else {
+			data.DiscoveryDevices[i].DeviceManagementIpAddress = types.StringNull()
+		}
 	}
 }
 
@@ -301,6 +422,15 @@ func (data *LANAutomation) isNull(ctx context.Context, res gjson.Result) bool {
 		return false
 	}
 	if !data.RedistributeIsisToBgp.IsNull() {
+		return false
+	}
+	if !data.DiscoveryLevel.IsNull() {
+		return false
+	}
+	if !data.DiscoveryTimeout.IsNull() {
+		return false
+	}
+	if len(data.DiscoveryDevices) > 0 {
 		return false
 	}
 	return true